requirements.txt
azure-functions==1.17.0
openai>=1.40.0
requests>=2.31.0
numpy==1.26.4
faiss-cpu==1.8.0
packaging